+++ README.md
@@ -7,7 +7,7 @@ Implement as much options below as you can:
 
 1. Render products in a table with id, name, category and an owner (user).
     - category should render its icon before the title;
-    - owner names should be colored with `has-text-link` for men and `has-text-danger` for women.
+    - owner names should be colored with `has-text-link` for men and `has-text-danger` for women. DONE
 1. Implement ability to filter products by an owner:
     - If a user is selected it should be highlighter with `is-active` class;
     - Show only products of a selected user;
@@ -33,7 +33,7 @@ Implement as much options below as you can:
 ## Instructions
 - Fork, clone and run `npm i`
 - fix the DEMO LINK below (use your github username and the repo name)
-  - [DEMO LINK](https://<your-account>.github.io/<your-repo-name>)
+  - [DEMO LINK](https://maksym-bobryk.github.io/<your-repo-name>)
 - implement tasks one by one (You can do it in the `App.jsx`)
 - `commit`, `push` and `deploy` after each task
 - Stop when the time is over (typically 2.5 hours from start)
